Fixed translation errors (#142) Fixed Could not authenticate against github.com (#130) Fixed some smaller bugs

+++ app/Console/Commands/Admin/Groups/DeleteGroupCommand.php
@@ -15,6 +15,12 @@ class DeleteGroupCommand extends Command
     {
 
         $group = GroupManager::findGroupByName($this->argument('name'));
+
+        if (!$group) {
+            $this->error('Group not found.');
+            return;
+        }
+
         $group->delete();
 
         $this->info('Group deleted successfully.');
